requirements.txt
google-cloud-aiplatform==1.0.1
python-binance==1.0.12
wget==3.2